% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic.R
\name{analisis_paparan}
\alias{analisis_paparan}
\title{Analisis bivariat seluruh variabel paparan}
\usage{
analisis_paparan(df, paparan, desain = "kohort")
}
\arguments{
\item{df}{Tabel hasil \code{\link[=bangun_dataset]{bangun_dataset()}}.}

\item{paparan}{\code{data.frame} kolom dan label paparan.}

\item{desain}{Desain studi.}
}
\value{
\code{data.frame} hasil analisis per variabel.
}
\description{
Analisis bivariat seluruh variabel paparan
}
