% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{baca_pemetaan}
\alias{baca_pemetaan}
\title{Membaca pemetaan variabel satu investigasi}
\usage{
baca_pemetaan(con, investigasi_id)
}
\arguments{
\item{con}{Koneksi basis data.}

\item{investigasi_id}{Id investigasi.}
}
\value{
Daftar berisi \code{pemetaan}, \code{konfigurasi}, dan \code{pembanding}.
}
\description{
Membaca pemetaan variabel satu investigasi
}
