% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{baca_submission}
\alias{baca_submission}
\title{Membaca seluruh submission satu investigasi}
\usage{
baca_submission(con, investigasi_id)
}
\arguments{
\item{con}{Koneksi basis data.}

\item{investigasi_id}{Id investigasi.}
}
\value{
\code{data.frame} submission.
}
\description{
Membaca seluruh submission satu investigasi
}
