% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{bagian_laporan}
\alias{bagian_laporan}
\title{Urutan bab laporan KLB}
\usage{
bagian_laporan()
}
\value{
\code{data.frame} dengan kolom \code{kunci} dan \code{judul}.
}
\description{
Mengikuti Panduan Penulisan Laporan KLB dengan format IMRAD.
}
