% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{boleh_ubah}
\alias{boleh_ubah}
\title{Apakah peran pengguna boleh mengubah data}
\usage{
boleh_ubah(peran)
}
\arguments{
\item{peran}{Peran pengguna.}
}
\value{
\code{TRUE} untuk admin dan analis.
}
\description{
Apakah peran pengguna boleh mengubah data
}
