% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{buka_db}
\alias{buka_db}
\title{Membuka koneksi basis data node}
\usage{
buka_db(peran = c("kabupaten", "provinsi"), berkas = NULL)
}
\arguments{
\item{peran}{\code{"kabupaten"} atau \code{"provinsi"}.}

\item{berkas}{Lokasi berkas SQLite. Bila \code{NULL} memakai \code{\link[=direktori_data]{direktori_data()}}.}
}
\value{
Koneksi DBI.
}
\description{
Membuka koneksi basis data node
}
