% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{catat_audit}
\alias{catat_audit}
\title{Mencatat aktivitas pengguna}
\usage{
catat_audit(con, pengguna_id, aksi, detail = NULL)
}
\arguments{
\item{con}{Koneksi basis data.}

\item{pengguna_id}{Id pengguna.}

\item{aksi}{Nama aksi.}

\item{detail}{Detail tambahan.}
}
\description{
Mencatat aktivitas pengguna
}
