% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{checklist_laporan}
\alias{checklist_laporan}
\title{Checklist kelengkapan laporan KLB}
\usage{
checklist_laporan()
}
\value{
\code{data.frame} butir checklist.
}
\description{
Disusun dari Checklist Laporan KLB Keracunan Makanan dan Panduan Penulisan
Laporan KLB, diselaraskan dengan format laporan pada Pedoman KLB 2020
Bab II bagian VIII.
}
