% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{daftarkan_node}
\alias{daftarkan_node}
\title{Mendaftarkan node kabupaten pada webapp provinsi}
\usage{
daftarkan_node(con, kode, nama, kabupaten)
}
\arguments{
\item{con}{Koneksi basis data provinsi.}

\item{kode}{Kode node.}

\item{nama}{Nama node.}

\item{kabupaten}{Nama kabupaten atau kota.}
}
\value{
Kunci API dalam bentuk teks.
}
\description{
Kunci API dibuat acak, disimpan sebagai hash, dan hanya dikembalikan satu
kali pada saat pembuatan.
}
