% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ekspor_markdown}
\alias{ekspor_markdown}
\title{Mengekspor laporan ke Markdown}
\usage{
ekspor_markdown(hasil, keterangan, naskah, berkas = NULL)
}
\arguments{
\item{hasil}{Objek \code{klb_hasil}.}

\item{keterangan}{Identitas investigasi.}

\item{naskah}{Daftar naskah bab dari \code{\link[=susun_draf_laporan]{susun_draf_laporan()}}.}

\item{berkas}{Lokasi berkas keluaran.}
}
\value{
Teks Markdown, tidak terlihat bila \code{berkas} diberikan.
}
\description{
Mengekspor laporan ke Markdown
}
