% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{ganti_sandi}
\alias{ganti_sandi}
\title{Mengganti kata sandi pengguna}
\usage{
ganti_sandi(con, pengguna_id, sandi_baru)
}
\arguments{
\item{con}{Koneksi basis data.}

\item{pengguna_id}{Id pengguna.}

\item{sandi_baru}{Kata sandi baru, minimal 8 karakter.}
}
\description{
Mengganti kata sandi pengguna
}
