% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{kelompokkan_umur}
\alias{kelompokkan_umur}
\title{Kelompok umur menurut batas yang ditentukan}
\usage{
kelompokkan_umur(umur, batas = c(0, 5, 13, 19, 46, 65))
}
\arguments{
\item{umur}{Vektor umur dalam tahun.}

\item{batas}{Vektor batas bawah kelompok, misalnya \code{c(0, 5, 13, 19, 46, 65)}.}
}
\value{
Faktor kelompok umur.
}
\description{
Kelompok umur menurut batas yang ditentukan
}
