% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klbfederated-package.R
\docType{package}
\name{klbfederated-package}
\alias{klbfederated}
\alias{klbfederated-package}
\title{klbfederated: Analisis Federated Kejadian Luar Biasa untuk Dinas Kesehatan}
\description{
Menyediakan dua aplikasi Shiny untuk penyelidikan kejadian luar biasa (KLB) dengan arsitektur federated. Node kabupaten atau kota menarik data investigasi dari KoboToolbox, menjalankan analisis epidemiologi sesuai Pedoman Penyelidikan dan Penanggulangan KLB Kementerian Kesehatan edisi revisi III tahun 2020, menyusun draf laporan, lalu mengirim ringkasan agregat ke dashboard provinsi. Data individu tidak pernah meninggalkan node kabupaten atau kota. Fungsi analisis juga dapat dipanggil langsung dari konsol R untuk pekerjaan di luar aplikasi.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/awatsiq/klbfederated}
  \item Report bugs at \url{https://github.com/awatsiq/klbfederated/issues}
}

}
\author{
\strong{Maintainer}: Ahmad Watsiq Maula \email{ahmad.watsiq.m@mail.ugm.ac.id}

Authors:
\itemize{
  \item Ahmad Watsiq Maula \email{ahmad.watsiq.m@mail.ugm.ac.id}
}

}
\keyword{internal}
