% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo_ambil_data}
\alias{kobo_ambil_data}
\title{Mengunduh seluruh submission satu proyek}
\usage{
kobo_ambil_data(base_url, token, uid, fields = NULL)
}
\arguments{
\item{base_url}{URL server KoboToolbox, misalnya \verb{https://kf.kobotoolbox.org}.}

\item{token}{Token API pengguna.}

\item{uid}{UID proyek KoboToolbox.}

\item{fields}{Definisi pertanyaan dari \code{\link[=kobo_definisi_form]{kobo_definisi_form()}}.}
}
\value{
\code{data.frame} submission.
}
\description{
Kolom \code{select_multiple} dimekarkan menjadi kolom biner sesuai definisi
formulir agar mudah dipetakan sebagai variabel gejala atau pangan.
}
