% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo_definisi_form}
\alias{kobo_definisi_form}
\title{Definisi pertanyaan pada formulir KoboToolbox}
\usage{
kobo_definisi_form(base_url, token, uid)
}
\arguments{
\item{base_url}{URL server KoboToolbox, misalnya \verb{https://kf.kobotoolbox.org}.}

\item{token}{Token API pengguna.}

\item{uid}{UID proyek KoboToolbox.}
}
\value{
Daftar berisi \code{nama} proyek dan \code{fields} (\code{data.frame}).
}
\description{
Pertanyaan bertipe \code{select_multiple} disertai daftar kolom hasil pemekaran
agar dapat dipetakan sebagai variabel biner.
}
