% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo_uji_koneksi}
\alias{kobo_uji_koneksi}
\title{Menguji koneksi ke KoboToolbox}
\usage{
kobo_uji_koneksi(base_url, token)
}
\arguments{
\item{base_url}{URL server KoboToolbox, misalnya \verb{https://kf.kobotoolbox.org}.}

\item{token}{Token API pengguna.}
}
\value{
Daftar berisi \code{username} dan \code{server}.
}
\description{
Menguji koneksi ke KoboToolbox
}
