% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{kriteria_klb}
\alias{kriteria_klb}
\title{Tujuh kriteria penetapan KLB}
\usage{
kriteria_klb()
}
\value{
\code{data.frame} dengan kolom \code{kode}, \code{uraian}, \code{otomatis}, dan \code{pembanding}.
}
\description{
Kriteria menurut Peraturan Menteri Kesehatan Nomor 1501 Tahun 2010,
sebagaimana dikutip pada Pedoman KLB 2020 Bab I bagian D butir 2.
}
