% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic.R
\name{nilai_kriteria_klb}
\alias{nilai_kriteria_klb}
\title{Penilaian tujuh kriteria KLB}
\usage{
nilai_kriteria_klb(pembanding, ringkasan, bins)
}
\arguments{
\item{pembanding}{Data pembanding surveilans, lihat \code{\link[=pembanding_kosong]{pembanding_kosong()}}.}

\item{ringkasan}{Daftar berisi \code{total_kasus} dan \code{cfr}.}

\item{bins}{Bins kurva epidemik.}
}
\value{
\code{data.frame} hasil penilaian tiap kriteria.
}
\description{
Penilaian tujuh kriteria KLB
}
