% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{normalkan_gejala}
\alias{normalkan_gejala}
\title{Menormalkan label gejala menjadi kunci gejala baku}
\usage{
normalkan_gejala(label)
}
\arguments{
\item{label}{Karakter label gejala.}
}
\value{
Kunci gejala baku atau \code{NA}.
}
\description{
Menormalkan label gejala menjadi kunci gejala baku
}
