% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{pembanding_kosong}
\alias{pembanding_kosong}
\title{Data pembanding kosong untuk penilaian kriteria KLB}
\usage{
pembanding_kosong()
}
\value{
Daftar bernama berisi nilai \code{NA}.
}
\description{
Data pembanding kosong untuk penilaian kriteria KLB
}
