% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{pengaturan}
\alias{pengaturan}
\title{Menyimpan atau membaca pengaturan}
\usage{
pengaturan(con, kunci, nilai = NULL)
}
\arguments{
\item{con}{Koneksi basis data.}

\item{kunci}{Nama pengaturan.}

\item{nilai}{Nilai baru. Bila \code{NULL}, fungsi membaca nilai.}
}
\value{
Nilai pengaturan.
}
\description{
Menyimpan atau membaca pengaturan
}
