% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{peran_wajib}
\alias{peran_wajib}
\title{Peran variabel yang wajib untuk satu jenis KLB}
\usage{
peran_wajib(jenis)
}
\arguments{
\item{jenis}{Jenis KLB: \code{"keracunan_pangan"}, \code{"penyakit_menular"}, atau \code{"pd3i"}.}
}
\value{
\code{data.frame} berisi peran wajib.
}
\description{
Peran variabel yang wajib untuk satu jenis KLB
}
