% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{periksa_login}
\alias{periksa_login}
\title{Memeriksa kredensial pengguna}
\usage{
periksa_login(con, username, sandi)
}
\arguments{
\item{con}{Koneksi basis data.}

\item{username}{Nama pengguna.}

\item{sandi}{Kata sandi.}
}
\value{
Daftar data pengguna bila berhasil, \code{NULL} bila gagal.
}
\description{
Memeriksa kredensial pengguna
}
