% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_attack_rate}
\alias{plot_attack_rate}
\title{Grafik attack rate menurut kategori}
\usage{
plot_attack_rate(tabel, judul = "Attack rate menurut kelompok")
}
\arguments{
\item{tabel}{\code{data.frame} hasil \code{\link[=tabel_attack_rate]{tabel_attack_rate()}}.}

\item{judul}{Judul gambar.}
}
\value{
Objek ggplot.
}
\description{
Grafik attack rate menurut kategori
}
