% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gejala}
\alias{plot_gejala}
\title{Grafik distribusi gejala}
\usage{
plot_gejala(gejala, judul = "Distribusi tanda dan gejala pada kasus")
}
\arguments{
\item{gejala}{\code{data.frame} distribusi gejala.}

\item{judul}{Judul gambar.}
}
\value{
Objek ggplot.
}
\description{
Grafik distribusi gejala
}
