% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_kurva_epidemik}
\alias{plot_kurva_epidemik}
\title{Kurva epidemik}
\usage{
plot_kurva_epidemik(hasil, judul = "Kurva epidemik KLB")
}
\arguments{
\item{hasil}{Objek \code{klb_hasil} dari \code{\link[=analisis_klb]{analisis_klb()}}.}

\item{judul}{Judul gambar.}
}
\value{
Objek ggplot.
}
\description{
Kurva epidemik
}
\examples{
contoh <- contoh_keracunan_pangan(n = 60)
hasil <- analisis_klb(contoh$data, contoh$pemetaan, contoh$konfigurasi)
plot_kurva_epidemik(hasil)
}
