% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_paparan}
\alias{plot_paparan}
\title{Grafik hasil analisis paparan}
\usage{
plot_paparan(paparan, judul = "Hasil analisis bivariat variabel paparan")
}
\arguments{
\item{paparan}{\code{data.frame} hasil \code{\link[=analisis_paparan]{analisis_paparan()}}.}

\item{judul}{Judul gambar.}
}
\value{
Objek ggplot.
}
\description{
Menampilkan estimasi RR atau OR beserta selang kepercayaan 95 persen pada
skala logaritmik, dengan garis acuan pada nilai satu.
}
