% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{putar_kunci_node}
\alias{putar_kunci_node}
\title{Memutar ulang kunci API satu node}
\usage{
putar_kunci_node(con, node_id)
}
\arguments{
\item{con}{Koneksi basis data provinsi.}

\item{node_id}{Id node.}
}
\value{
Kunci API baru.
}
\description{
Memutar ulang kunci API satu node
}
