% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{simpan_pemetaan}
\alias{simpan_pemetaan}
\title{Menyimpan pemetaan variabel dan konfigurasi analisis}
\usage{
simpan_pemetaan(
  con,
  investigasi_id,
  pemetaan,
  konfigurasi,
  pembanding = pembanding_kosong()
)
}
\arguments{
\item{con}{Koneksi basis data.}

\item{investigasi_id}{Id investigasi.}

\item{pemetaan}{Daftar pemetaan peran variabel.}

\item{konfigurasi}{Konfigurasi analisis.}

\item{pembanding}{Data pembanding kriteria KLB.}
}
\description{
Menyimpan pemetaan variabel dan konfigurasi analisis
}
