% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{tabel_laporan}
\alias{tabel_laporan}
\title{Menyusun tabel laporan bernomor}
\usage{
tabel_laporan(hasil, keterangan)
}
\arguments{
\item{hasil}{Objek \code{klb_hasil}.}

\item{keterangan}{Identitas investigasi.}
}
\value{
Daftar tabel, masing-masing berisi \code{nomor}, \code{judul}, \code{data}, dan \code{catatan}.
}
\description{
Susunan tabel mengikuti Checklist Laporan KLB Keracunan Makanan.
}
