% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tema_klb}
\alias{tema_klb}
\title{Tema grafik untuk laporan KLB}
\usage{
tema_klb(dasar = 11)
}
\arguments{
\item{dasar}{Ukuran huruf dasar.}
}
\value{
Objek tema ggplot2.
}
\description{
Tema minimalis tanpa garis kisi tebal agar gambar siap dipakai pada laporan
cetak dan naskah publikasi.
}
