% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-helpers.R
\name{versi_aset}
\alias{versi_aset}
\title{Penanda versi aset antarmuka}
\usage{
versi_aset()
}
\value{
Teks versi package.
}
\description{
Ditempelkan pada URL berkas CSS dan JavaScript agar peramban mengambil versi
terbaru setelah package diperbarui.
}
